Update storybook icon
Storybook has updated their icon a while back, this follows that.
Note that I kept the "s" white. I know you might prefer not to but to me it doesn't look like a storybook logo with a "hollow" s on dark background.
It's in line with e.g. current gatsby logo in this repo which also has solid white inside.