Add support for git-hooks as console file icon (#1138)
The husky tool is great for git hooks and those files are shell scripts. Since `sh` is denoted for the console icon we can adapt those for git hooks because they are shell scripts overall. On a new git init browse `.git/hooks` folder and those all are shell scripts although there is no extension `sh`.

Ref: https://github.com/wpeverest/everest-forms/tree/develop/.husky