fix: filter clone language icons for fileIcons.png
The Autohotkey2 language icon (1f95e80a) was the first language icon to
be a clone of another icon. This led to "ahk2" showing as a missing
reference in the fileIcons.png file for the latest few releases.

This fix adds a filter to remove clone language icons when collecting
the list of icons to generate the fileIcons.png file.