fix: update Blender logo to upstream version
It was noticed in the Blender community chat that the logo used in Gitea appeared to come from the material-extensions source, but was not using the correct logo.

The guidelines and design for the logo can be found at:

  https://www.blender.org/about/logo/

This patch attempts to honor the spirit of the replaced logo, while following the "slanted" look of the upstream logo.